requirements.txt
# Django and REST framework
Django>=5.2.1,<5.3.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.0.0,<4.1.0
django-filter>=23.2,<23.3
django-oauth-toolkit>=2.3.0,<2.4.0
djangorestframework-simplejwt>=5.2.2,<5.3.0

# API Documentation
drf-yasg>=1.21.5,<1.22.0
drf-spectacular>=0.26.2,<0.27.0

# Database
psycopg2-binary>=2.9.6,<2.10.0  

# Caching and Message Queue
redis>=4.5.5,<4.6.0
django-redis>=5.2.0,<5.3.0
celery>=5.3.1,<5.4.0
flower>=2.0.0,<2.1.0 

# Security
cryptography>=40.0.2,<40.1.0
django-cors-headers>=4.0.0,<4.1.0

# Utilities
Pillow>=9.5.0,<9.6.0  
python-dateutil>=2.8.2,<2.9.0
pytz>=2023.3
requests>=2.30.0,<2.31.0

# Testing
pytest>=7.3.1,<7.4.0
pytest-django>=4.5.2,<4.6.0
coverage>=7.2.5,<7.3.0

# Development
black>=23.3.0,<23.4.0  
flake8>=6.0.0,<6.1.0  
isort>=5.12.0,<5.13.0  
